+++ transport/go17.go
@@ -36,6 +36,7 @@ package transport
 
 import (
 	"net"
+	"net/http"
 
 	"golang.org/x/net/context"
 )
@@ -43,4 +44,12 @@ import (
 // dialContext connects to the address on the named network.
 func dialContext(ctx context.Context, network, address string) (net.Conn, error) {
 	return (&net.Dialer{}).DialContext(ctx, network, address)
+}
+
+func sendRequest(ctx context.Context, req *http.Request, conn net.Conn) error {
+	req = req.WithContext(ctx)
+	if err := req.Write(conn); err != nil {
+		return err
+	}
+	return nil
 }
